+++ src/main/java/chess/dao/dto/TurnDto.java
@@ -1,4 +1,4 @@
-package chess.web.dto;
+package chess.dao.dto;
 
 import chess.domain.state.StateType;
 import java.util.Objects;
