+++ src/test/java/chess/dao/TurnDaoTest.java
@@ -3,7 +3,7 @@ package chess.dao;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.state.StateType;
-import chess.web.dto.TurnDto;
+import chess.dao.dto.TurnDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
